requirements.txt
sanic~=21.6.2
python-dotenv~=0.19.0
pytest~=6.2.5
pydantic~=1.8.2